Refacto identities overview
